% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizedMF.R
\name{cv.gmf}
\alias{cv.gmf}
\title{CV for generalized MF}
\usage{
cv.gmf(x, ks, family = c("gaussian", "binomial", "poisson", "multinomial"),
  folds = 5, quiet = TRUE, ...)
}
\arguments{
\item{x}{matrix of either binary, count, or continuous data}

\item{ks}{the different dimensions \code{k} to try}

\item{family}{exponential family distribution of data}

\item{folds}{if \code{folds} is a scalar, then it is the number of folds. If
it is a vector, it should be the same length as the number of rows in \code{x}}

\item{quiet}{logical; whether the function should display progress}

\item{...}{Additional arguments passed to generalizedMF}
}
\value{
A matrix of the CV deviance with \code{k} in rows
}
\description{
Run cross validation on dimension for generalized MF
}
\examples{
# construct a low rank matrix in the logit scale
rows = 100
cols = 10
set.seed(1)
mat_logit = outer(rnorm(rows), rnorm(cols))

# generate a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0

\dontrun{
deviances = cv.gmf(mat, ks = 1:9, family = "binomial")
plot(deviances)
}
}

