% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_family_functions.R
\name{exp_fam_deviance}
\alias{exp_fam_deviance}
\title{Calculate exponential family deviance}
\usage{
exp_fam_deviance(x, theta, family, weights = 1)
}
\arguments{
\item{x}{a vector or matrix of data}

\item{theta}{natural parameters. Must be same dimensions as \code{x}}

\item{family}{exponential family distribution of data}

\item{weights}{weights of datapoints. Must be a scalar or the same dimensions as the data}
}
\description{
Calculates the deviance of data assuming it comes from an exponential family
}

