% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_family_harmonium.R
\name{exponential_family_harmonium}
\alias{exponential_family_harmonium}
\title{Exponential Family Harmoniums}
\usage{
exponential_family_harmonium(x, k = 2, family = c("gaussian", "binomial",
  "poisson"), family_hidden = c("gaussian", "binomial", "poisson"),
  cd_iters = 1, learning_rate = 0.001, max_iters = 100,
  rms_prop = FALSE, quiet = TRUE, random_start = TRUE, start_W, mu,
  main_effects = TRUE)
}
\arguments{
\item{x}{matrix of either binary, count, or continuous data}

\item{k}{dimension (number of hidden units)}

\item{family}{exponential family distribution of data}

\item{family_hidden}{exponential family distribution of hidden units}

\item{cd_iters}{number of iterations for contrastive divergence (CD) at each iteration.
It should be a vector of two integers, which is the range of CD interations that the
algorithm will perform from the beginning until the end, linearly interpolated}

\item{learning_rate}{learning rate used for gradient descent}

\item{max_iters}{maximum number of iterations}

\item{quiet}{logical; whether the calculation should give feedback}

\item{random_start}{whether to randomly initialize \code{W}}

\item{start_W}{initial value for \code{W}}

\item{mu}{specific value for \code{mu}, the mean (bias) vector of \code{x}}

\item{main_effects}{logical; whether to include main effects (bias terms) in the model}
}
\value{
An S3 object of class \code{efh} which is a list with the
following components:
\item{mu}{the main effects (bias terms) for dimensionality reduction}
\item{hidden_bias}{the bias for the hidden units (currently hard coded to 0)}
\item{W}{the \code{d}x\code{k}-dimentional matrix with the loadings}
\item{family}{the exponential family of the data}
\item{family_hidden}{the exponential family of the hidden units}
\item{iters}{number of iterations required for convergence}
\item{loss_trace}{the trace of the average deviance of the algorithm.
   Should be non-increasing}
\item{prop_deviance_expl}{the proportion of deviance explained by this model.
   If \code{main_effects = TRUE}, the null model is just the main effects, otherwise
   the null model estimates 0 for all natural parameters.}
}
\description{
Welling et al. (2005)'s Exponential Family Harmoniums
}
\examples{
rows = 100
cols = 10
set.seed(1)
mat_np = outer(rnorm(rows), rnorm(cols))

# generate a count matrix and binary response
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)
mat[1, 1] <- NA

modp = exponential_family_harmonium(mat, k = 2, family = "poisson", quiet = FALSE, learning_rate = 0.001, rms_prop = F, max_iters = 250)
gmf = generalizedMF(mat, k = 1, family = "poisson", quiet = FALSE)

}
\references{
Welling, Max, Michal Rosen-Zvi, and Geoffrey E. Hinton. "Exponential family
harmoniums with an application to information retrieval." Advances in neural
information processing systems. 2005.
}

