% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizedMF.R
\name{generalizedMF}
\alias{generalizedMF}
\title{Exponential Family Matrix Factorization}
\usage{
generalizedMF(x, k = 2, family = c("gaussian", "binomial", "poisson"),
  quiet = TRUE, max_iters = 1000, conv_criteria = 1e-05,
  random_start = FALSE, start_A, start_B, mu, main_effects = TRUE,
  method = c("als", "svd"))
}
\arguments{
\item{x}{matrix of either binary, proportions, count, or continuous data}

\item{k}{dimension}

\item{family}{exponential family distribution of data}

\item{quiet}{logical; whether the calculation should give feedback}

\item{max_iters}{maximum number of iterations}

\item{conv_criteria}{convergence criteria}

\item{random_start}{whether to randomly initialize \code{A} and \code{B}}

\item{start_A}{initial value for \code{A}}

\item{start_B}{initial value for \code{B}}

\item{mu}{specific value for \code{mu}, the mean vector of \code{x}}

\item{main_effects}{logical; whether to include main effects in the model}

\item{method}{which algorithm to use. \code{"als"} uses alternating least squares.
It has the benefit of majozing row-wise and column-wise for each of the updates.
\code{"svd"} uses singular value decomposition (similar to de Leeuw, 2006). It has to
a more gereral majorization, which may not work well for heterogeneous matrices.}
}
\value{
An S3 object of class \code{gsmf} which is a list with the
following components:
\item{mu}{the main effects for dimensionality reduction}
\item{A}{the \code{n}x\code{k}-dimentional matrix with the scores}
\item{B}{the \code{d}x\code{k}-dimentional matrix with the loadings}
\item{beta}{the \code{k + 1} length vector of the coefficients}
\item{family_x}{the exponential family of covariates}
\item{family_y}{the exponential family of response}
\item{iters}{number of iterations required for convergence}
\item{loss_trace}{the trace of the average deviance of the algorithm.
   Should be non-increasing}
\item{prop_deviance_expl}{the proportion of deviance explained by this model.
   If \code{main_effects = TRUE}, the null model is just the main effects, otherwise
   the null model estimates 0 for all natural parameters.}
}
\description{
Collins et al. (2001)'s Exponential Family PCA
}
\examples{
rows = 100
cols = 10
set.seed(1)
mat_np = outer(rnorm(rows), rnorm(cols))

# generate a count matrix and binary response
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)

mod = generalizedMF(mat, k = 1, family = "poisson", quiet = FALSE)

}
\references{
de Leeuw, Jan, 2006. Principal component analysis of binary data
by iterated singular value decomposition. Computational Statistics & Data Analysis
50 (1), 21--39.

Collins, M., Dasgupta, S., & Schapire, R. E., 2001. A generalization of principal
components analysis to the exponential family. In NIPS, 617--624.
}

