% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizedPCA-package.R
\docType{package}
\name{generalizedPCA-package}
\alias{generalizedPCA-package}
\title{generalizedPCA-package}
\description{
Generalized dimensionality reduction by extending PCA
}
\author{
Andrew J. Landgraf
}
\keyword{package}

