% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{generalizedPCA}
\alias{generalizedPCA}
\title{Generalized Principal Component Analysis}
\usage{
generalizedPCA(x, k = 2, M = 4, family = c("gaussian", "binomial",
  "poisson", "multinomial"), weights, quiet = TRUE, majorizer = c("row",
  "all"), use_irlba = FALSE, max_iters = 1000, conv_criteria = 1e-05,
  random_start = FALSE, start_U, start_mu, main_effects = TRUE, validation,
  val_weights)
}
\arguments{
\item{x}{matrix of either binary, proportions, count, or continuous data}

\item{k}{number of principal components to return}

\item{M}{value to approximate the saturated model}

\item{family}{exponential family distribution of data}

\item{weights}{an optional matrix of the same size as the \code{x} with data weights}

\item{quiet}{logical; whether the calculation should give feedback}

\item{majorizer}{how to majorize the deviance. \code{"row"} gives
tighter majorization, but may take longer to calculate each iteration.
\code{"all"} may be faster per iteration, but take more iterations}

\item{use_irlba}{logical; if \code{TRUE}, the function uses the irlba package
to more quickly calculate the eigen-decomposition}

\item{max_iters}{number of maximum iterations}

\item{conv_criteria}{convergence criteria. The difference between average deviance
in successive iterations}

\item{random_start}{logical; whether to randomly inititalize the parameters. If \code{FALSE},
function will use an eigen-decomposition as starting value}

\item{start_U}{starting value for the orthogonal matrix}

\item{start_mu}{starting value for mu. Only used if \code{main_effects = TRUE}}

\item{main_effects}{logical; whether to include main effects in the model}
}
\value{
An S3 object of class \code{gpca} which is a list with the
following components:
\item{mu}{the main effects}
\item{U}{a \code{k}-dimentional orthonormal matrix with the loadings}
\item{PCs}{the princial components}
\item{M}{the parameter inputed}
\item{family}{the exponential family used}
\item{iters}{number of iterations required for convergence}
\item{loss_trace}{the trace of the average negative log likelihood of the algorithm.
   Should be non-increasing}
\item{prop_deviance_expl}{the proportion of deviance explained by this model.
   If \code{main_effects = TRUE}, the null model is just the main effects, otherwise
   the null model estimates 0 for all natural parameters.}
}
\description{
Dimension reduction for exponential family data by extending Pearson's
PCA formulation
}
\examples{
# construct a low rank matrix in the natural parameter space
rows = 100
cols = 10
set.seed(1)
mat_np = outer(rnorm(rows), rnorm(cols))

# generate a count matrix
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)

# run Poisson PCA on it
gpca = generalizedPCA(mat, k = 1, M = 4, family = "poisson")
}

