% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexGeneralizedPCA.R
\name{plot.cgpca}
\alias{plot.cgpca}
\title{Plot convex generalized PCA}
\usage{
\method{plot}{cgpca}(x, type = c("trace", "loadings", "scores"), ...)
}
\arguments{
\item{x}{convex generalized PCA object}

\item{type}{the type of plot \code{type = "trace"} plots the algorithms progress by
iteration, \code{type = "loadings"} plots the first 2 PC loadings,
\code{type = "scores"} plots the first 2 PC scores}

\item{...}{Additional arguments}
}
\description{
Plots the results of a convex generalized PCA
}
\examples{
# construct a low rank matrix in the logit scale
rows = 100
cols = 10
set.seed(1)
mat_logit = outer(rnorm(rows), rnorm(cols))

# generate a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0

# run convex generalized PCA on it
cgpca = convexGeneralizedPCA(mat, k = 1, M = 4, family = "binomial")

\dontrun{
plot(cgpca)
}
}

