% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{plot.cv.gpca}
\alias{plot.cv.gpca}
\title{Plot CV for generalized PCA}
\usage{
\method{plot}{cv.gpca}(object, ...)
}
\arguments{
\item{object}{a \code{cv.gpca} object}

\item{...}{Additional arguments}
}
\description{
Plot cross validation results generalized PCA
}

