% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_family_harmonium.R
\name{plot.efh}
\alias{plot.efh}
\title{Plot exponential family harmonium}
\usage{
\method{plot}{efh}(x, type = c("trace"), ...)
}
\arguments{
\item{x}{EFH object}

\item{type}{the type of plot \code{type = "trace"} plots the algorithms progress by
iteration}

\item{...}{Additional arguments}
}
\description{
Plots the results of a EFH
}
\examples{
# construct a low rank matrix in the logit scale
rows = 100
cols = 10
set.seed(1)
mat_logit = outer(rnorm(rows), rnorm(cols))

# generate a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0

# run logistic SVD on it
efh = exponential_family_harmonium(mat, k = 2, family = "binomial", family_hidden = "binomial")

\dontrun{
plot(efh)
}
}

