% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{plot.gpca}
\alias{plot.gpca}
\title{Plot generalized PCA}
\usage{
\method{plot}{gpca}(object, type = c("trace", "loadings", "scores"), ...)
}
\arguments{
\item{object}{generalized PCA object}

\item{type}{the type of plot \code{type = "trace"} plots the algorithms progress by
iteration, \code{type = "loadings"} plots the first 2 principal component
loadings, \code{type = "scores"} plots the loadings first 2 principal component scores}

\item{...}{Additional arguments}
}
\description{
Plots the results of a generalized PCA
}
\examples{
# construct a low rank matrix in the natural parameter scale
rows = 100
cols = 10
set.seed(1)
mat_np = outer(rnorm(rows), rnorm(cols))

# generate a count matrix
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)

# run logistic PCA on it
gpca = generalizedPCA(mat, k = 2, M = 4, family = "poisson")

\dontrun{
plot(gpca)
}
}

