% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizedMF.R
\name{predict.gmf}
\alias{predict.gmf}
\title{Predict generalized PCA scores or reconstruction on new data}
\usage{
\method{predict}{gmf}(object, newdata, type = c("PCs", "link", "response"),
  quiet = TRUE, max_iters = 1000, conv_criteria = 1e-05, start_A, ...)
}
\arguments{
\item{object}{generalized MF object}

\item{newdata}{matrix of the same exponential family as covariates in \code{object}.
If missing, will use the data that \code{object} was fit on}

\item{type}{the type of fitting required.
\code{type = "PCs"} gives matrix of principal components of \code{x},
\code{type = "link"} gives a matrix on the natural parameter scale, and
\code{type = "response"} gives a matrix on the response scale}

\item{quiet}{logical; whether the calculation should show progress}

\item{max_iters}{maximum number of iterations}

\item{conv_criteria}{convergence criteria}

\item{start_A}{initial value for \code{A}}

\item{...}{Additional arguments}
}
\description{
Predict generalized PCA scores or reconstruction on new data
}
\examples{
# construct a low rank matrices in the natural parameter space
rows = 100
cols = 10
set.seed(1)
loadings = rnorm(cols)
mat_np = outer(rnorm(rows), rnorm(cols))
mat_np_new = outer(rnorm(rows), loadings)

# generate a count matrices
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)
mat_new = matrix(rpois(rows * cols, c(exp(mat_np_new))), rows, cols)

# run Poisson PCA on it
gmf = generalizedMF(mat, k = 1, family = "poisson")

A = predict(gmf, mat_new)

}

