% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{predict.gpca}
\alias{predict.gpca}
\title{Predict generalized PCA scores or reconstruction on new data}
\usage{
\method{predict}{gpca}(object, newdata, type = c("PCs", "link", "response"),
  ...)
}
\arguments{
\item{object}{generalized PCA object}

\item{newdata}{matrix of the same exponential family as in \code{object}.
If missing, will use the data that \code{object} was fit on}

\item{type}{the type of fitting required. \code{type = "PCs"} gives the PC scores,
\code{type = "link"} gives matrix on the natural parameter scale and
\code{type = "response"} gives matrix on the response scale}

\item{...}{Additional arguments}
}
\description{
Predict generalized PCA scores or reconstruction on new data
}
\examples{
# construct a low rank matrices in the natural parameter space
rows = 100
cols = 10
set.seed(1)
loadings = rnorm(cols)
mat_np = outer(rnorm(rows), rnorm(cols))
mat_np_new = outer(rnorm(rows), loadings)

# generate a count matrices
mat = matrix(rpois(rows * cols, c(exp(mat_np))), rows, cols)
mat_new = matrix(rpois(rows * cols, c(exp(mat_np_new))), rows, cols)

# run Poisson PCA on it
gpca = generalizedPCA(mat, k = 1, M = 4, family = "poisson")

PCs = predict(gpca, mat_new)
}

