% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{print.gpca}
\alias{print.gpca}
\title{Print generalized PCA object}
\usage{
\method{print}{gpca}(x, ...)
}
\arguments{
\item{x}{generalized PCA object}

\item{...}{Additional arguments}
}
\description{
Print generalized PCA object
}

