% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexGeneralizedPCA.R
\name{project.Fantope}
\alias{project.Fantope}
\title{Project onto the Fantope}
\usage{
project.Fantope(x, k, partial_decomp = FALSE)
}
\arguments{
\item{x}{a symmetric matrix}

\item{k}{the rank of the Fantope desired}

\item{partial_decomp}{logical; if \code{TRUE}, the function uses the RSpectra package
to more quickly calculate the eigen-decomposition. When the number of columns is small,
the approximation may be less accurate and slower}
}
\value{
\item{H}{a rank \code{k} Fantope matrix}
\item{U}{a \code{k}-dimentional orthonormal matrix with the first \code{k} eigenvectors of \code{H}}
}
\description{
Project a symmetric matrix onto the convex set of the rank k Fantope
}

