% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{cv.gpca}
\alias{cv.gpca}
\title{CV for generalized PCA}
\usage{
cv.gpca(x, ks, Ms = seq(2, 10, by = 2), family = c("gaussian", "binomial",
  "poisson"), folds = 5, quiet = TRUE, ...)
}
\arguments{
\item{x}{matrix of either binary, count, or continuous data}

\item{ks}{the different dimensions \code{k} to try}

\item{Ms}{the different approximations to the saturated model \code{M} to try}

\item{family}{exponential family distribution of data}

\item{folds}{if \code{folds} is a scalar, then it is the number of folds. If
it is a vector, it should be the same length as the number of rows in \code{x}}

\item{quiet}{logical; whether the function should display progress}

\item{...}{Additional arguments passed to \code{generalizedPCA}}
}
\value{
A matrix of the CV log likelihood with \code{k} in rows and
 \code{M} in columns
}
\description{
Run cross validation on dimension and \code{M} for generalized PCA
}

