% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{fitted.gpca}
\alias{fitted.gpca}
\title{Fitted values using generalized PCA}
\usage{
\method{fitted}{gpca}(object, type = c("link", "response"), ...)
}
\arguments{
\item{object}{generalized PCA object}

\item{type}{the type of fitting required. \code{type = "link"} gives output on the natural
parameter scale and \code{type = "response"} gives output on the response scale}

\item{...}{Additional arguments}
}
\description{
Fit a lower dimentional representation of the exponential family matrix using generalized PCA
}

