% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generalizedPCA.R
\name{predict.gpca}
\alias{predict.gpca}
\title{Predict generalized PCA scores or reconstruction on new data}
\usage{
\method{predict}{gpca}(object, newdata, type = c("PCs", "link", "response"),
  ...)
}
\arguments{
\item{object}{generalized PCA object}

\item{newdata}{matrix of the same exponential family as in \code{object}.
If missing, will use the data that \code{object} was fit on}

\item{type}{the type of fitting required. \code{type = "PCs"} gives the PC scores,
\code{type = "link"} gives matrix on the natural parameter scale and
\code{type = "response"} gives matrix on the response scale}

\item{...}{Additional arguments}
}
\description{
Predict generalized PCA scores or reconstruction on new data
}

